Editing the header and making new components along with sending off the correct dates for the API
